// types/contentTypes.d.ts
export type targetType = "shonen" | "shojo" | "seinen" | "josei" | "kodomo";
export type categoryType =
  | "manga"
  | "comics"
  | "manhwa"
  | "manhua"
  | "light_novel";

export type SelectOptionType = {
  label: string;
  value: string;
};

export type statusType = "on_going" | "hiatus" | "finished";